Makefile - external library support
added possibility, for adding linker flags, so that binary external libs
will link correctly.

One unsolved issue:
For some reason, the linker doesn't recognize the symbols in the static
external lib. In order to compile, the static lib needs to be unpacked,
and the Makefile searches for all object files, as a workaround.